precentacion de proyecto

+++ src/screens/createroutes/WriteRoute.js
@@ -5,48 +5,16 @@ import { stylesMain } from "../../stylesMain";
 import { collection, addDoc, serverTimestamp, deleteDoc, doc } from "firebase/firestore";
 import { getAuth } from "firebase/auth";
 import { FireBase_DB } from '../../../data/FirebaseConfig';
+import Icon from "react-native-vector-icons/Ionicons";
 
 export default function WriteRoute(props) {
     const routeId = props.route.params.routeId;
-    const [userLocation, setUserLocation] = useState(null);
     const [pointCounter, setPointCounter] = useState(0);
 
-    const startLocationUpdates = async () => {
-        try {
-            let { status } = await Location.requestForegroundPermissionsAsync();
-            if (status !== 'granted') {
-                console.log('Permission to access location was denied');
-                return;
-            }
-
-            await Location.watchPositionAsync(
-                {
-                    accuracy: Location.Accuracy.High,
-                    distanceInterval: 2
-                },
-                (location) => {
-                    setUserLocation(location.coords);
-                }
-            );
-        } catch (error) {
-            console.error('Error al iniciar la actualización de ubicación:', error);
-        }
-    };
-
-    useEffect(() => {
-        startLocationUpdates();
-    }, []);
-
     const saveStep = async (description) => {
-        if (!userLocation) {
-            console.log("No se pudo obtener la ubicación del usuario.");
-            return;
-        }
 
         const stepData = {
             description,
-            latitude: userLocation.latitude,
-            longitude: userLocation.longitude,
             creationtime: serverTimestamp()
         };
 
@@ -76,29 +44,33 @@ export default function WriteRoute(props) {
     };
 
     const cancelRoute = async (id) => {
-        try {
-            const routeDocRef = doc(FireBase_DB, 'usuarios', '1Rf7KYV0jvN8CAfzZIEF', 'Rutas', id);
-            await deleteDoc(routeDocRef)
 
-            console.log("Todo a salido de maravilla")
-        } catch (error) {
-            console.log("No se a podido cancelar la accion", error)
+        const auth = getAuth();
+        const user = auth.currentUser;
+
+        if (user) {
+            const uid = user.uid;
+            try {
+                const routeDocRef = doc(FireBase_DB, 'usuarios', uid, 'Rutas', id);
+                await deleteDoc(routeDocRef)
+                
+                console.log("Todo a salido de maravilla")
+            } catch (error) {
+                console.log("No se a podido cancelar la accion", error)
+            }
+        } else {
+            console.log("No hay usuario autentificado")
         }
     };
 
-    console.log("Estos son los puntos que llevo: ", pointCounter)
-
     return (
         <View style={stylesMain.main}>
-            <Text>
-                En este apartado podrás dibujar una ruta que después podrás escuchar.
-            </Text>
 
             <View style={styles.containerBtns}>
                 {[
-                    { title: "Línea Recta", style: styles.btnRecta },
-                    { title: "Vuelta a la izquierda", style: styles.btnIzquierda },
-                    { title: "Vuelta a la derecha", style: styles.btnDerecha },
+                    { title: "Línea Recta", style: styles.btnRecta, icon: 'arrow-up-outline' },
+                    { title: "Vuelta a la izquierda", style: styles.btnIzquierda, icon: 'arrow-back-outline'},
+                    { title: "Vuelta a la derecha", style: styles.btnDerecha, icon: 'arrow-forward-outline'},
                     { title: "Dejar Precionado Para Cancelar y Presiona Para Terminar", style: styles.btnCancelar, navigate: true, cancel: true },
                 ].map((btn, index) => (
                     <Pressable
@@ -125,7 +97,15 @@ export default function WriteRoute(props) {
                             { opacity: pressed ? 0.6 : 1.0 }
                         ]}
                     >
-                        <Text style={styles.title}>{btn.title}</Text>
+                        {btn.icon ? (
+                            <Icon
+                                name={btn.icon}
+                                size={100}
+                                color={"#1E2022"}
+                            />
+                        ) : (
+                            <Text style={styles.title}>{btn.title}</Text>
+                        )}
                     </Pressable>
                 ))}
             </View>
@@ -140,37 +120,47 @@ const styles = StyleSheet.create({
         height: 550
     },
     btnRecta: {
-        backgroundColor: 'red',
+        backgroundColor: 'gray',
         width: 360,
         height: 150,
         justifyContent: 'center',
-        alignItems: 'center'
+        alignItems: 'center',
+        borderWidth: 2,
+        borderColor: 'black'
     },
     btnIzquierda: {
-        backgroundColor: 'yellow',
+        backgroundColor: 'gray',
         width: 180,
         height: 250,
         justifyContent: 'center',
-        alignItems: 'center'
+        alignItems: 'center',
+        borderWidth: 2,
+        borderColor: 'black'
     },
     btnDerecha: {
         position: 'absolute',
         right: 0,
         bottom: 150,
-        backgroundColor: 'green',
+        backgroundColor: 'gray',
         width: 180,
         height: 250,
         justifyContent: 'center',
-        alignItems: 'center'
+        alignItems: 'center',
+        borderWidth: 2,
+        borderColor: 'black'
     },
     btnCancelar: {
         backgroundColor: 'gray',
         width: 360,
         height: 150,
         justifyContent: 'center',
-        alignItems: 'center'
+        alignItems: 'center',
+        borderWidth: 2,
+        borderColor: 'black'
     },
     title: {
-        fontSize: 25
+        fontSize: 18,
+        color: 'white',
+        textAlign: 'center'
     }
 });
